[Refactor] 패키지 구조 재정의 (Config 및 DTO 이동)
- KakaoClientConfig: config -> spot/config 패키지로 이동
- SpotResponseDto: spot/dto -> spot/dto/response 패키지로 이동
